Rename AgentFactory projects and assemblies to Orchestration
