+++ src/FoundationaLLM.sln
@@ -47,9 +47,9 @@ Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "PromptHubAPI", "python\Prom
 EndProject
 Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "AgentHubAPI", "python\AgentHubAPI\AgentHubAPI.pyproj", "{865A944E-76C4-4DA9-B367-6D1CAB66ABCC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AgentFactoryAPI", "dotnet\AgentFactoryAPI\AgentFactoryAPI.csproj", "{87CBD379-8F88-4C45-B7B1-595B5BAE734D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OrchestrationAPI", "dotnet\OrchestrationAPI\OrchestrationAPI.csproj", "{87CBD379-8F88-4C45-B7B1-595B5BAE734D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AgentFactory", "dotnet\AgentFactory\AgentFactory.csproj", "{59C6DEDC-F60D-4EB9-A2E5-793EC42F3895}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orchestration", "dotnet\Orchestration\Orchestration.csproj", "{59C6DEDC-F60D-4EB9-A2E5-793EC42F3895}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestUtils", "..\tests\dotnet\TestUtils\TestUtils.csproj", "{7F7CA6E9-F1F0-403D-B72D-2AC134F64CDE}"
 EndProject
@@ -91,9 +91,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Configuration", "dotnet\Con
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataSource", "dotnet\DataSource\DataSource.csproj", "{9E243B01-0DA5-4E7B-9539-7E5BF87106EE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Authorization", "dotnet\Authorization\Authorization.csproj", "{10D4F412-0F01-485F-957E-9EA8F16AED10}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Authorization", "dotnet\Authorization\Authorization.csproj", "{10D4F412-0F01-485F-957E-9EA8F16AED10}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AuthorizationAPI", "dotnet\AuthorizationAPI\AuthorizationAPI.csproj", "{63D7285C-0C40-49E9-BB4D-8AD2E0A29E62}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AuthorizationAPI", "dotnet\AuthorizationAPI\AuthorizationAPI.csproj", "{63D7285C-0C40-49E9-BB4D-8AD2E0A29E62}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
