--- src/dotnet/AgentFactory/Models/ConfigurationOptions/LangChainServiceSettings.cs
@@ -1,18 +1,0 @@
-﻿using FoundationaLLM.AgentFactory.Core.Models.ConfigurationOptions;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using FoundationaLLM.Common.Models.Configuration.API;
-
-namespace FoundationaLLM.AgentFactory.Models.ConfigurationOptions
-{
-    /// <summary>
-    /// The settings configuration options for a LangChain Service
-    /// </summary>
-    public class LangChainServiceSettings : SettingsBase
-    {
-        
-    }
-}
