+++ src/dotnet/Orchestration/Models/ConfigurationOptions/AzureAIDirectServiceSettings.cs
@@ -1,6 +1,6 @@
 ﻿using FoundationaLLM.Common.Models.Configuration.API;
 
-namespace FoundationaLLM.AgentFactory.Core.Models.ConfigurationOptions
+namespace FoundationaLLM.Orchestration.Core.Models.ConfigurationOptions
 {
     /// <summary>
     /// The settings configuration options for an Azure AI direct service.
