+++ src/dotnet/Orchestration/Models/ConfigurationOptions/LangChainServiceSettings.cs
@@ -1,0 +1,12 @@
+﻿using FoundationaLLM.Common.Models.Configuration.API;
+
+namespace FoundationaLLM.Orchestration.Core.Models.ConfigurationOptions
+{
+    /// <summary>
+    /// The settings configuration options for a LangChain Service
+    /// </summary>
+    public class LangChainServiceSettings : SettingsBase
+    {
+        
+    }
+}
