+++ src/dotnet/Orchestration/Models/ConfigurationOptions/SemanticKernelServiceSettings.cs
@@ -1,7 +1,6 @@
-﻿using FoundationaLLM.AgentFactory.Core.Models.ConfigurationOptions;
-using FoundationaLLM.Common.Models.Configuration.API;
+﻿using FoundationaLLM.Common.Models.Configuration.API;
 
-namespace FoundationaLLM.AgentFactory.Models.ConfigurationOptions
+namespace FoundationaLLM.Orchestration.Core.Models.ConfigurationOptions
 {
     /// <summary>
     /// The settings configuration options for a SemanticKernalService
