+++ src/dotnet/Orchestration/Orchestration/KnowledgeManagementOrchestration.cs
@@ -1,10 +1,10 @@
-﻿using FoundationaLLM.AgentFactory.Interfaces;
-using FoundationaLLM.Common.Interfaces;
+﻿using FoundationaLLM.Common.Interfaces;
 using FoundationaLLM.Common.Models.Agents;
 using FoundationaLLM.Common.Models.Orchestration;
+using FoundationaLLM.Orchestration.Core.Interfaces;
 using Microsoft.Extensions.Logging;
 
-namespace FoundationaLLM.AgentFactory.Core.Orchestration
+namespace FoundationaLLM.Orchestration.Core.Orchestration
 {
     /// <summary>
     /// Knowledge Management orchestration.
@@ -12,14 +12,14 @@ namespace FoundationaLLM.AgentFactory.Core.Orchestration
     /// <remarks>
     /// Constructor for default agent.
     /// </remarks>
-    /// <param name="agent">The <see cref="InternalContextAgent"/> agent.</param>
+    /// <param name="agent">The <see cref="KnowledgeManagementAgent"/> agent.</param>
     /// <param name="callContext">The call context of the request being handled.</param>
     /// <param name="orchestrationService"></param>
     /// <param name="promptHubService"></param>
     /// <param name="dataSourceHubService"></param>
     /// <param name="logger">The logger used for logging.</param>
-    public class InternalContextOrchestration(
-        InternalContextAgent agent,
+    public class KnowledgeManagementOrchestration(
+        KnowledgeManagementAgent agent,
         ICallContext callContext,
         ILLMOrchestrationService orchestrationService,
         IPromptHubAPIService promptHubService,
@@ -28,7 +28,7 @@ namespace FoundationaLLM.AgentFactory.Core.Orchestration
     {
         private readonly ICallContext _callContext = callContext;
         private readonly ILogger<OrchestrationBase> _logger = logger;
-        private readonly InternalContextAgent _agent = agent;
+        private readonly KnowledgeManagementAgent _agent = agent;
 
         /// <inheritdoc/>
         public override Task Configure(CompletionRequest completionRequest) =>
@@ -50,6 +50,7 @@ namespace FoundationaLLM.AgentFactory.Core.Orchestration
             {
                 Completion = result.Completion!,
                 UserPrompt = completionRequest.UserPrompt!,
+                Citations = result.Citations,
                 FullPrompt = result.FullPrompt,
                 PromptTemplate = result.PromptTemplate,
                 AgentName = result.AgentName,
