+++ src/dotnet/Orchestration/Services/LangChainService.cs
@@ -1,15 +1,15 @@
-﻿using FoundationaLLM.Common.Models.Orchestration;
-using FoundationaLLM.AgentFactory.Interfaces;
-using FoundationaLLM.AgentFactory.Models.ConfigurationOptions;
-using FoundationaLLM.Common.Interfaces;
-using FoundationaLLM.Common.Settings;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using System.Text;
+﻿using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using FoundationaLLM.Common.Interfaces;
+using FoundationaLLM.Common.Models.Orchestration;
+using FoundationaLLM.Common.Settings;
+using FoundationaLLM.Orchestration.Core.Interfaces;
+using FoundationaLLM.Orchestration.Core.Models.ConfigurationOptions;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 
-namespace FoundationaLLM.AgentFactory.Services
+namespace FoundationaLLM.Orchestration.Core.Services
 {
     /// <summary>
     /// The LangChain orchestration service.
