+++ src/dotnet/OrchestrationAPI/Dockerfile
@@ -1,0 +1,34 @@
+#See https://aka.ms/customizecontainer to learn how to customize your debug container and how Visual Studio uses this Dockerfile to build your images for faster debugging.
+FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS base
+ARG FOUNDATIONALLM_VERSION
+WORKDIR /app
+EXPOSE 80
+
+ENV ASPNETCORE_URLS=http://+:80
+ENV ASPNETCORE_ENVIRONMENT=Production
+ENV FOUNDATIONALLM_VERSION=${FOUNDATIONALLM_VERSION:-0.0.0}
+
+FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
+WORKDIR /src
+COPY . .
+
+RUN dotnet restore "dotnet/Common/Common.csproj"
+RUN dotnet build "dotnet/Common/Common.csproj" -c Release -o /app/build
+
+RUN dotnet restore "dotnet/AgentFactory/AgentFactory.csproj"
+RUN dotnet build "dotnet/AgentFactory/AgentFactory.csproj" -c Release -o /app/build
+
+RUN dotnet restore "dotnet/AgentFactoryAPI/AgentFactoryAPI.csproj"
+RUN dotnet build "dotnet/AgentFactoryAPI/AgentFactoryAPI.csproj" -c Release -o /app/build
+
+FROM build AS publish
+RUN dotnet publish "dotnet/Common/Common.csproj" -c Release -o /app/publish /p:UseAppHost=false
+RUN dotnet publish "dotnet/AgentFactory/AgentFactory.csproj" -c Release -o /app/publish /p:UseAppHost=false
+RUN dotnet publish "dotnet/AgentFactoryAPI/AgentFactoryAPI.csproj" -c Release -o /app/publish /p:UseAppHost=false
+
+FROM base AS final
+WORKDIR /app
+COPY --from=publish /app/publish .
+ENTRYPOINT ["dotnet", "FoundationaLLM.AgentFactory.API.dll"]
+LABEL org.opencontainers.image.source=https://github.com/solliancenet/foundationallm
+LABEL ai.foundationallm.version=${FOUNDATIONALLM_VERSION:-0.0.0}
