+++ tests/dotnet/AgentFactory.Tests/Models/ConfigurationOptions/SemanticKernelServiceSettingsTests.cs
@@ -1,4 +1,4 @@
-﻿using FoundationaLLM.AgentFactory.Models.ConfigurationOptions;
+﻿using FoundationaLLM.Orchestration.Core.Models.ConfigurationOptions;
 
 namespace FoundationaLLM.AgentFactory.Tests.Models.ConfigurationOptions
 {
