+++ tests/dotnet/AgentFactory.Tests/Orchestration/KnowledgeManagementOrchestrationTests.cs
@@ -1,8 +1,8 @@
-﻿using FoundationaLLM.AgentFactory.Core.Orchestration;
-using FoundationaLLM.AgentFactory.Interfaces;
-using FoundationaLLM.Common.Interfaces;
+﻿using FoundationaLLM.Common.Interfaces;
 using FoundationaLLM.Common.Models.Agents;
 using FoundationaLLM.Common.Models.Orchestration;
+using FoundationaLLM.Orchestration.Core.Interfaces;
+using FoundationaLLM.Orchestration.Core.Orchestration;
 
 namespace FoundationaLLM.AgentFactory.Tests.Orchestration
 {
