+++ tests/dotnet/AgentFactory.Tests/Services/SemanticKernelServiceTests.cs
@@ -1,8 +1,9 @@
-﻿using FoundationaLLM.AgentFactory.Models.ConfigurationOptions;
-using FoundationaLLM.AgentFactory.Services;
-using FoundationaLLM.Common.Interfaces;
+﻿using FoundationaLLM.Common.Interfaces;
 using FoundationaLLM.Common.Models.Agents;
 using FoundationaLLM.Common.Models.Orchestration;
+using FoundationaLLM.Orchestration.Core.Models.ConfigurationOptions;
+using FoundationaLLM.Orchestration.Core.Services;
+
 namespace FoundationaLLM.AgentFactory.Tests.Services
 {
     public class SemanticKernelServiceTests
